+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to the
 
 ## Unreleased
 
+- Fixed various issues in the requirements document and added missing specifications for check execution behaviour.
+
 ## 0.1.1 - 2025-10-08
 
 - Changed default commit range from `HEAD` to `HEAD~1..HEAD`.