+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to the
 - Fixed check failure report entries having the wrong keys for exit codes and missing the expected exit code.
 - Changed `on-repository` and `on-commit` configuration keys to be optional.
 - Fixed per-file checks ignoring the configured `expected-exit` code.
+- Fixed invalid configuration file not causing a non-zero exit code.
 
 ## 0.1.1 - 2025-10-08